Fixed and updated tests
